split functions up into smaller files
